+++ test/functional-tests/node-transformers/converting-transformers/split-string-transformer/fixtures/string-with-emoji-1.js
@@ -1,1 +1,2 @@
-'ab👋🏼cd';
+var test = 'ab👋🏼cd';
+test;